requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0

# Data loading / transforms
pillow>=9.5.0
opencv-python>=4.8.0

tqdm>=4.66.0

# Optional logging
wandb>=0.16.0
